+++ spectest/qbft/run_msg.go
@@ -1,0 +1,88 @@
+package qbft
+
+import (
+	"bytes"
+	"errors"
+	"testing"
+
+	specqbft "github.com/bloxapp/ssv-spec/qbft"
+	spectests "github.com/bloxapp/ssv-spec/qbft/spectest/tests"
+	"github.com/stretchr/testify/require"
+)
+
+func runMsgSpecTest(t *testing.T, test *spectests.MsgSpecTest) {
+	var lastErr error
+
+	for i, messageBytes := range test.EncodedMessages {
+		m := &specqbft.SignedMessage{}
+		if err := m.Decode(messageBytes); err != nil {
+			lastErr = err
+		}
+
+		if len(test.ExpectedRoots) > 0 {
+			r, err := m.GetRoot()
+			if err != nil {
+				lastErr = err
+			}
+
+			if !bytes.Equal(test.ExpectedRoots[i], r) {
+				t.Fail()
+			}
+		}
+	}
+
+	for i, msg := range test.Messages {
+		if err := msg.Validate(); err != nil {
+			lastErr = err
+		}
+
+		switch msg.Message.MsgType {
+		case specqbft.RoundChangeMsgType:
+			rc := specqbft.RoundChangeData{}
+			if err := rc.Decode(msg.Message.Data); err != nil {
+				lastErr = err
+			}
+			if err := validateRoundChangeData(rc); err != nil {
+				lastErr = err
+			}
+		}
+
+		if len(test.Messages) > 0 {
+			r1, err := msg.Encode()
+			if err != nil {
+				lastErr = err
+			}
+
+			r2, err := test.Messages[i].Encode()
+			if err != nil {
+				lastErr = err
+			}
+			if !bytes.Equal(r2, r1) {
+				t.Fail()
+			}
+		}
+	}
+
+	if len(test.ExpectedError) != 0 {
+		require.EqualError(t, lastErr, test.ExpectedError)
+	} else {
+		require.NoError(t, lastErr)
+	}
+}
+
+func validateRoundChangeData(d specqbft.RoundChangeData) error {
+	if isRoundChangeDataPrepared(d) {
+		if len(d.PreparedValue) == 0 {
+			return errors.New("round change prepared value invalid")
+		}
+		if len(d.RoundChangeJustification) == 0 {
+			return errors.New("round change justification invalid")
+		}
+	}
+
+	return nil
+}
+
+func isRoundChangeDataPrepared(d specqbft.RoundChangeData) bool {
+	return d.PreparedRound != 0 || len(d.PreparedValue) != 0
+}
