+++ spectest/qbft/run_msg_processing.go
@@ -1,0 +1,224 @@
+package qbft
+
+import (
+	"context"
+	"encoding/hex"
+	"fmt"
+	"testing"
+
+	specqbft "github.com/bloxapp/ssv-spec/qbft"
+	spectests "github.com/bloxapp/ssv-spec/qbft/spectest/tests"
+	spectypes "github.com/bloxapp/ssv-spec/types"
+	"github.com/bloxapp/ssv-spec/types/testingutils"
+	"github.com/stretchr/testify/require"
+	"go.uber.org/zap"
+	"go.uber.org/zap/zapcore"
+
+	forksprotocol "github.com/bloxapp/ssv/protocol/forks"
+	beaconprotocol "github.com/bloxapp/ssv/protocol/v1/blockchain/beacon"
+	protocolp2p "github.com/bloxapp/ssv/protocol/v1/p2p"
+	qbftprotocol "github.com/bloxapp/ssv/protocol/v1/qbft"
+	forksfactory "github.com/bloxapp/ssv/protocol/v1/qbft/controller/forks/factory"
+	"github.com/bloxapp/ssv/protocol/v1/qbft/instance"
+	"github.com/bloxapp/ssv/protocol/v1/qbft/instance/leader/constant"
+	"github.com/bloxapp/ssv/protocol/v1/qbft/instance/msgcont"
+	qbftstorage "github.com/bloxapp/ssv/protocol/v1/qbft/storage"
+	"github.com/bloxapp/ssv/protocol/v1/types"
+	"github.com/bloxapp/ssv/protocol/v1/validator"
+	"github.com/bloxapp/ssv/storage"
+	"github.com/bloxapp/ssv/storage/basedb"
+	"github.com/bloxapp/ssv/utils/logex"
+)
+
+func runMsgProcessingSpecTest(t *testing.T, test *spectests.MsgProcessingSpecTest) {
+	ctx := context.TODO()
+	logger := logex.Build(test.Name, zapcore.DebugLevel, nil)
+
+	forkVersion := forksprotocol.GenesisForkVersion
+	pi, _ := protocolp2p.GenPeerID()
+	p2pNet := protocolp2p.NewMockNetwork(logger, pi, 10)
+	beacon := validator.NewTestBeacon(t)
+
+	var keysSet *testingutils.TestKeySet
+	switch len(test.Pre.State.Share.Committee) {
+	case 4:
+		keysSet = testingutils.Testing4SharesSet()
+	case 7:
+		keysSet = testingutils.Testing7SharesSet()
+	case 10:
+		keysSet = testingutils.Testing10SharesSet()
+	case 13:
+		keysSet = testingutils.Testing13SharesSet()
+	default:
+		t.Error("unknown key set length")
+	}
+
+	db, err := storage.GetStorageFactory(basedb.Options{
+		Type:   "badger-memory",
+		Logger: logger,
+		Ctx:    ctx,
+	})
+	require.NoError(t, err)
+
+	qbftStorage := qbftstorage.NewQBFTStore(db, logger, spectypes.BNRoleAttester.String())
+
+	share := &beaconprotocol.Share{
+		NodeID:    1,
+		PublicKey: keysSet.ValidatorPK,
+		Committee: make(map[spectypes.OperatorID]*beaconprotocol.Node),
+	}
+
+	var mappedCommittee []*spectypes.Operator
+	for i := range test.Pre.State.Share.Committee {
+		operatorID := spectypes.OperatorID(i) + 1
+		require.NoError(t, beacon.AddShare(keysSet.Shares[operatorID]))
+
+		pk := keysSet.Shares[operatorID].GetPublicKey().Serialize()
+		share.Committee[spectypes.OperatorID(i)+1] = &beaconprotocol.Node{
+			IbftID: uint64(i) + 1,
+			Pk:     pk,
+		}
+
+		share.OperatorIds = append(share.OperatorIds, uint64(i)+1)
+
+		mappedCommittee = append(mappedCommittee, &spectypes.Operator{
+			OperatorID: operatorID,
+			PubKey:     pk,
+		})
+	}
+
+	mappedShare := &spectypes.Share{
+		OperatorID:      share.NodeID,
+		ValidatorPubKey: share.PublicKey.Serialize(),
+		SharePubKey:     share.Committee[share.NodeID].Pk,
+		Committee:       mappedCommittee,
+		Quorum:          keysSet.Threshold,
+		PartialQuorum:   keysSet.PartialThreshold,
+		DomainType:      types.GetDefaultDomain(),
+		Graffiti:        nil,
+	}
+
+	qbftInstance := newQbftInstance(logger, qbftStorage, p2pNet, beacon, share, mappedShare, test.Pre.StartValue, forkVersion)
+	qbftInstance.Init()
+	qbftInstance.State().InputValue.Store(test.Pre.StartValue)
+	qbftInstance.State().Round.Store(test.Pre.State.Round)
+	qbftInstance.State().Height.Store(test.Pre.State.Height)
+	qbftInstance.State().ProposalAcceptedForCurrentRound.Store(test.Pre.State.ProposalAcceptedForCurrentRound)
+
+	var lastErr error
+	for _, msg := range test.InputMessages {
+		if _, err := qbftInstance.ProcessMsg(msg); err != nil {
+			lastErr = err
+		}
+	}
+
+	mappedInstance := new(specqbft.Instance)
+	if qbftInstance != nil {
+		preparedValue := qbftInstance.State().GetPreparedValue()
+		if len(preparedValue) == 0 {
+			preparedValue = nil
+		}
+		round := qbftInstance.State().GetRound()
+
+		_, decidedErr := qbftInstance.CommittedAggregatedMsg()
+		var decidedValue []byte
+		if decidedErr == nil && len(test.InputMessages) != 0 {
+			decidedValue = test.InputMessages[0].Message.Identifier
+		}
+
+		mappedInstance.State = &specqbft.State{
+			Share:                           mappedShare,
+			ID:                              test.Pre.State.ID,
+			Round:                           round,
+			Height:                          qbftInstance.State().GetHeight(),
+			LastPreparedRound:               qbftInstance.State().GetPreparedRound(),
+			LastPreparedValue:               preparedValue,
+			ProposalAcceptedForCurrentRound: qbftInstance.State().GetProposalAcceptedForCurrentRound(),
+			Decided:                         decidedErr == nil,
+			DecidedValue:                    decidedValue,
+			ProposeContainer:                convertToSpecContainer(t, qbftInstance.Containers()[specqbft.ProposalMsgType]),
+			PrepareContainer:                convertToSpecContainer(t, qbftInstance.Containers()[specqbft.PrepareMsgType]),
+			CommitContainer:                 convertToSpecContainer(t, qbftInstance.Containers()[specqbft.CommitMsgType]),
+			RoundChangeContainer:            convertToSpecContainer(t, qbftInstance.Containers()[specqbft.RoundChangeMsgType]),
+		}
+
+		mappedInstance.StartValue = qbftInstance.State().GetInputValue()
+	}
+
+	if len(test.ExpectedError) == 0 {
+		require.NoError(t, lastErr)
+	} else if mappedExpectedError, ok := errorsMap[test.ExpectedError]; ok {
+		require.Equal(t, mappedExpectedError, lastErr.Error())
+	} else {
+		require.EqualError(t, lastErr, test.ExpectedError)
+	}
+
+	mappedRoot, err := mappedInstance.State.GetRoot()
+	require.NoError(t, err)
+	require.Equal(t, test.PostRoot, hex.EncodeToString(mappedRoot))
+
+	type BroadcastMessagesGetter interface {
+		GetBroadcastMessages() []spectypes.SSVMessage
+	}
+
+	outputMessages := p2pNet.(BroadcastMessagesGetter).GetBroadcastMessages()
+
+	require.Equal(t, len(test.OutputMessages), len(outputMessages))
+
+	for i, outputMessage := range outputMessages {
+		r1, _ := test.OutputMessages[i].GetRoot()
+
+		msg2 := &specqbft.SignedMessage{}
+		require.NoError(t, msg2.Decode(outputMessage.Data))
+
+		r2, _ := msg2.GetRoot()
+		require.EqualValues(t, r1, r2, fmt.Sprintf("output msg %d roots not equal", i))
+	}
+
+	db.Close()
+}
+
+func newQbftInstance(logger *zap.Logger, qbftStorage qbftstorage.QBFTStore, net protocolp2p.MockNetwork, beacon *validator.TestBeacon, share *beaconprotocol.Share, mappedShare *spectypes.Share, identifier []byte, forkVersion forksprotocol.ForkVersion) instance.Instancer {
+	const height = 0
+	fork := forksfactory.NewFork(forkVersion)
+
+	newInstance := instance.NewInstance(&instance.Options{
+		Logger:           logger,
+		ValidatorShare:   share,
+		Network:          net,
+		Config:           qbftprotocol.DefaultConsensusParams(),
+		Identifier:       identifier[:],
+		Height:           height,
+		RequireMinPeers:  false,
+		Fork:             forkWithValueCheck{fork.InstanceFork()},
+		SSVSigner:        beacon.KeyManager,
+		ChangeRoundStore: qbftStorage,
+	})
+	//newInstance.(*instance.Instance).LeaderSelector = roundRobinLeaderSelector{newInstance.State(), mappedShare}
+	// TODO(nkryuchkov): replace when ready
+	newInstance.(*instance.Instance).LeaderSelector = &constant.Constant{
+		LeaderIndex: 0,
+		OperatorIDs: share.OperatorIds,
+	}
+	return newInstance
+}
+
+func convertToSpecContainer(t *testing.T, container msgcont.MessageContainer) *specqbft.MsgContainer {
+	c := specqbft.NewMsgContainer()
+	container.AllMessaged(func(round specqbft.Round, msg *specqbft.SignedMessage) {
+		ok, err := c.AddIfDoesntExist(&specqbft.SignedMessage{
+			Signature: msg.Signature,
+			Signers:   msg.Signers,
+			Message: &specqbft.Message{
+				MsgType:    msg.Message.MsgType,
+				Height:     msg.Message.Height,
+				Round:      msg.Message.Round,
+				Identifier: msg.Message.Identifier,
+				Data:       msg.Message.Data,
+			},
+		})
+		require.NoError(t, err)
+		require.True(t, ok)
+	})
+	return c
+}
